Multi-threading Producer consumer example added
This example only works sequentially, but prevents any type of race conditions or data inconsistency. In order to make it parallel, you need to create mutexes that corresponds to each buffer index. So if your buffer has a 20 size index, you need at least 20+20 (20 for producer and 20 for consumer), or less.