Parallel version of producer consumer code added
This is a solution for the practice final exam. Producers and consumers interacts with a 20 sized buffer array. Producers and consumers finds the available index in the array and performs their task. However, this program is not deadlock free. If producers are too much than the consumers an when there is no place left to produce something, they will run in infinite loop. This scenario similarly applies for the consumers as well. To fix the issue a simple control can be added, which can be timed, or based on active consumer threads.