Dinning Philosopher Problem and Solution added
In the dinningPhilosopher.cpp code the deadlock problem and example has been given. In the dinningPhilosopherSolution.cpp the deadlock free solution is given by only adding an additional semaphore with a size 4, instead of 5. This allows one thread to be out of the critical section. Thereby, we guarentee that we have at least one additional resource (chopstick) available to keep thing running.